[Validator] Allow opt-out of EmailValidator deprecation when using Validation::createValidatorBuilder()
