+++ src/Symfony/Component/Validator/Tests/Fixtures/DummyConstraint.php
@@ -1,0 +1,22 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Validator\Tests\Fixtures;
+
+use Symfony\Component\Validator\Constraint;
+
+class DummyConstraint extends Constraint
+{
+    public function validatedBy(): string
+    {
+        return DummyConstraintValidator::class;
+    }
+}
